chore: renames NonMenuDropdown to Popover
This commit also renames the old Popover component to something more precise - PopoverBox
