+++ packages/index.ts
@@ -53,7 +53,6 @@ export {
   DropdownMenuItemAvatar,
   DropdownMenuItemAppearances
 } from "./dropdownMenu";
-export { NonMenuDropdown as Dropdown } from "./nonMenuDropdown";
 export {
   EmptyState,
   EmptyStateWithGraphic,
@@ -81,6 +80,7 @@ export {
   FullscreenModal
 } from "./modal";
 export { PageHeader, PageHeaderTabs } from "./pageheader";
+export { Popover } from "./popover";
 export { ProgressBar } from "./progressbar";
 export { PromoBanner, PromoInline } from "./promo";
 export { SegmentedControl, SegmentedControlButton } from "./segmentedControl";
