--- packages/nonMenuDropdown/index.ts
@@ -1,1 +1,0 @@
-export { default as NonMenuDropdown } from "./NonMenuDropdown";
