+++ packages/popover/index.ts
@@ -1,2 +1,1 @@
 export { default as Popover } from "./components/Popover";
-export { default as PopoverListItem } from "./components/PopoverListItem";
