+++ packages/popover/popover.stories.tsx
@@ -2,24 +2,24 @@ import * as React from "react";
 import { storiesOf } from "@storybook/react";
 import { select } from "@storybook/addon-knobs";
 import { withReadme } from "storybook-readme";
-import { NonMenuDropdown } from ".";
+import { Popover } from "../";
 import { Direction } from "../dropdownable/components/Dropdownable";
 import { PrimaryDropdownButton } from "../button";
 
 const readme = require("./README.md");
 
-storiesOf("Overlays|NonMenuDropdown", module)
+storiesOf("Overlays|Popover", module)
   .addDecorator(withReadme([readme]))
   .add("default", () => (
-    <NonMenuDropdown
-      trigger={<PrimaryDropdownButton>open dropdown</PrimaryDropdownButton>}
+    <Popover
+      trigger={<PrimaryDropdownButton>open popover</PrimaryDropdownButton>}
     >
       <div>dropdown content</div>
-    </NonMenuDropdown>
+    </Popover>
   ))
   .add("with focusable content", () => (
-    <NonMenuDropdown
-      trigger={<PrimaryDropdownButton>open dropdown</PrimaryDropdownButton>}
+    <Popover
+      trigger={<PrimaryDropdownButton>open popover</PrimaryDropdownButton>}
       maxWidth={300}
     >
       <div>
@@ -30,15 +30,15 @@ storiesOf("Overlays|NonMenuDropdown", module)
         </p>
         <p>Focus should not leave the dropdown until it is closed</p>
       </div>
-    </NonMenuDropdown>
+    </Popover>
   ))
   .add("initialIsOpen", () => (
-    <NonMenuDropdown
+    <Popover
       initialIsOpen={true}
-      trigger={<PrimaryDropdownButton>open dropdown</PrimaryDropdownButton>}
+      trigger={<PrimaryDropdownButton>open popover</PrimaryDropdownButton>}
     >
       <div>dropdown content</div>
-    </NonMenuDropdown>
+    </Popover>
   ))
   .add("with custom direction", () => {
     const options = {
@@ -59,18 +59,18 @@ storiesOf("Overlays|NonMenuDropdown", module)
     }
 
     return (
-      <NonMenuDropdown
-        trigger={<PrimaryDropdownButton>open dropdown</PrimaryDropdownButton>}
+      <Popover
+        trigger={<PrimaryDropdownButton>open popover</PrimaryDropdownButton>}
         preferredDirections={[Direction[getKeyByValue(knobDirection)]]}
       >
         Use the knobs to change dropdown alignment
-      </NonMenuDropdown>
+      </Popover>
     );
   })
   .add("with maxHeight", () => (
-    <NonMenuDropdown
+    <Popover
       maxHeight={80}
-      trigger={<PrimaryDropdownButton>open dropdown</PrimaryDropdownButton>}
+      trigger={<PrimaryDropdownButton>open popover</PrimaryDropdownButton>}
     >
       <div>dropdown content</div>
       <p>
@@ -81,12 +81,12 @@ storiesOf("Overlays|NonMenuDropdown", module)
         centuries, but also the leap into electronic typesetting, remaining
         essentially unchanged.
       </p>
-    </NonMenuDropdown>
+    </Popover>
   ))
   .add("with maxWidth", () => (
-    <NonMenuDropdown
+    <Popover
       maxWidth={300}
-      trigger={<PrimaryDropdownButton>open dropdown</PrimaryDropdownButton>}
+      trigger={<PrimaryDropdownButton>open popover</PrimaryDropdownButton>}
     >
       <div>dropdown content</div>
       <p>
@@ -97,5 +97,5 @@ storiesOf("Overlays|NonMenuDropdown", module)
         centuries, but also the leap into electronic typesetting, remaining
         essentially unchanged.
       </p>
-    </NonMenuDropdown>
+    </Popover>
   ));
